SMTChecker: Upgrade CVC4 to cvc5
Instead of compiling `solc` itself with CVC4 support, it is now enough
to have `cvc5` executable on PATH when running the compiler.
Instead of using API of CVC4, we now use SMT-LIB2 interface.
That means we write the queries to temporary SMT-LIB2 files and call
the solver process directly to run on the file.