SMTChecker: Allow using multiple external solvers in one analysis
We are using SMTCommand inside UniversalCallback to call external
solvers on queries produced my our engines.

Previous mechanism set the external solver once during initialization
and it was not possible to change it later. This meant, that it would
not be possible to use, e.g., Eldarica and cvc5 at the same time.

Here we move the proper setup for SMTCommand just before we call it.
This setup is customized by subclasses of (CHC)SmtLib2Interface, which
call corresponding external solvers.